+++ quantify_core/visualization/SI_utilities.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 import re
 import string
-from typing import Optional, Tuple, Union
+from typing import List, Optional, Tuple, Union
 import warnings
 
 import lmfit
@@ -21,6 +21,85 @@ golden_mean = (np.sqrt(5) - 1.0) / 2.0  # Aesthetic ratio
 single_col_figsize = (3.39, golden_mean * 3.39)
 double_col_figsize = (6.9, golden_mean * 6.9)
 thesis_col_figsize = (12.2 / 2.54, golden_mean * 12.2 / 2.54)
+
+
+def _get_scale_factor_and_offset_and_prefix(
+    ticks: List[float], unit: str | None = None, precision: int = 4
+) -> Tuple[float, float, str]:
+    """Return a convenient scale factor, offset and SI prefix based on the tick values.
+
+    This function uses the :func:`~.SI_prefix_and_scale_factor` function to determine a
+    scale factor such that the distance between ticks is in the range [0.1, 100.0), plus
+    the corresponding scaled SI unit (e.g. 'mT', 'kV'), deduced from the input unit, to
+    represent the tick values in those scaled units. In addition, an offset is
+    calculated such that the maximum absolute tick value is less than 10^precision.
+
+    .. admonition:: Example
+
+        .. jupyter-execute::
+            :hide-code:
+
+            from quantify_core.visualization.SI_utilities import (
+                _get_scale_factor_and_offset_and_prefix
+            )
+
+        .. jupyter-execute::
+
+            _get_scale_factor_and_offset_and_prefix(
+                ticks=[2100000, 2100100, 2100200],
+                unit="Hz",
+                precision=4
+            )
+            # (0.001, 21000, 'kHz')
+
+    Parameters
+    ----------
+    ticks : List[float]
+        A list of axis tick values.
+    unit : str or None, optional
+        The unit of the tick values.
+    precision : int, optional
+        The maximum amount of digits to display as tick labels.
+
+    Returns
+    -------
+    scale_factor : float
+        The scale factor to multiply the tick values with.
+    offset : float
+        The offset to subtract from the tick values.
+    unit : str
+        The unit including the SI prefix.
+    """
+    max_v, min_v = max(ticks), min(ticks)
+    resolution = (max_v - min_v) / len(ticks)
+    scale_factor, unit = SI_prefix_and_scale_factor(val=resolution * 10, unit=unit)
+    signed_max = max_v if abs(max_v) > abs(min_v) else min_v
+    factor = pow(10, precision - 1)
+    offset = int(signed_max * scale_factor / factor) * factor
+    return scale_factor, offset, unit
+
+
+def _set_offset_string(
+    formatter: matplotlib.ticker.Formatter, offset: float, unit: str
+) -> None:
+    """Set the offset string of the Formatter to a conveniently scaled offset.
+
+    This function scales the given offset and unit using
+    :func:`~.SI_prefix_and_scale_factor`, and sets the offset string of the Formatter to
+    the scaled offset value.
+
+    Parameters
+    ----------
+    formatter : matplotlib.ticker.Formatter
+        The matplotlib Formatter.
+    offset : float
+        The value to scale and display.
+    unit : str
+        The unit of the value.
+    """
+    offset_scale, offset_unit = SI_prefix_and_scale_factor(offset, unit)
+    disp_offset = offset * offset_scale
+    formatter.set_offset_string(f"{disp_offset:+g} {offset_unit}")
 
 
 def set_xlabel(
@@ -42,8 +121,9 @@ def set_xlabel(
     """
     if isinstance(label, plt.Axes):
         warnings.warn(
-            "Passing axis as a first argument is deprecated and will be removed in quantify-core >= 0.10.0."
-            " Please use the new syntax set_xlabel(label, unit = None, axis = None)",
+            "Passing axis as a first argument is deprecated and will be removed "
+            "in quantify-core >= 0.10.0. Please use the new syntax "
+            "set_xlabel(label, unit = None, axis = None)",
             FutureWarning,
             stacklevel=2,
         )
@@ -54,10 +134,17 @@ def set_xlabel(
 
     if unit:
         xticks = axis.get_xticks()
-        scale_factor, unit = SI_prefix_and_scale_factor(val=max(abs(xticks)), unit=unit)
-        formatter = matplotlib.ticker.FuncFormatter(
-            lambda x, pos: f"{x * scale_factor:.4g}"
+        precision = 4
+        scale_factor, offset, unit = _get_scale_factor_and_offset_and_prefix(
+            xticks, unit, precision
         )
+
+        formatter = matplotlib.ticker.FuncFormatter(
+            lambda x, pos: f"{x * scale_factor - offset:.{precision}g}"
+        )
+
+        if offset != 0:
+            _set_offset_string(formatter, offset, unit)
 
         axis.xaxis.set_major_formatter(formatter)
         axis.set_xlabel(label + f" [{unit}]", **kw)
@@ -97,10 +184,17 @@ def set_ylabel(
 
     if unit:
         yticks = axis.get_yticks()
-        scale_factor, unit = SI_prefix_and_scale_factor(val=max(abs(yticks)), unit=unit)
-        formatter = matplotlib.ticker.FuncFormatter(
-            lambda x, pos: f"{x * scale_factor:.6g}"
+        precision = 6
+        scale_factor, offset, unit = _get_scale_factor_and_offset_and_prefix(
+            yticks, unit, precision=precision
         )
+
+        formatter = matplotlib.ticker.FuncFormatter(
+            lambda x, pos: f"{x * scale_factor - offset:.{precision}g}"
+        )
+
+        if offset != 0:
+            _set_offset_string(formatter, offset, unit)
 
         axis.yaxis.set_major_formatter(formatter)
 
@@ -129,10 +223,18 @@ def set_cbarlabel(
     """
     if unit:
         zticks = cbar.get_ticks()
-        scale_factor, unit = SI_prefix_and_scale_factor(val=max(abs(zticks)), unit=unit)
-        formatter = matplotlib.ticker.FuncFormatter(
-            lambda x, pos: f"{x * scale_factor:.6g}"
+        precision = 6
+        scale_factor, offset, unit = _get_scale_factor_and_offset_and_prefix(
+            zticks, unit, precision=precision
         )
+
+        formatter = matplotlib.ticker.FuncFormatter(
+            lambda x, pos: f"{x * scale_factor - offset:.{precision}g}"
+        )
+
+        if offset != 0:
+            _set_offset_string(formatter, offset, unit)
+
         cbar.ax.yaxis.set_major_formatter(formatter)
         cbar.set_label(label + f" [{unit}]")